% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_testing.R
\name{.check_test_token}
\alias{.check_test_token}
\title{Check Test PAT}
\usage{
.check_test_token()
}
\value{
named vector of TRUE or FALSE as to whether the token is configured
and can access the test repos.
}
\description{
We use a fine-grained GitHub Personal Access Token in our auth testing. This
function is memoised and checks that the token has the correct permissions
to the various test repositories.
}
\details{
It should have the following token permissions:
\itemize{
\item Contents (read/write)
\item Metadata (read)
}

on the following repositories
\itemize{
\item tanho63/piggyback
\item tanho63/piggyback-tests
\item tanho63/piggyback-private (a private repository)
}
}
\keyword{internal}
