% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.gh_api_url}
\alias{.gh_api_url}
\title{GitHub API URL}
\usage{
.gh_api_url()
}
\value{
string: API base url
}
\description{
Reads environment variable GITHUB_API_URL to determine base URL of API. Same
as gh package. Defaults to \verb{https://api.github.com}.
}
\seealso{
\url{https://gh.r-lib.org/#environment-variables}
}
