% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_caching.R
\name{.pb_cache_clear}
\alias{.pb_cache_clear}
\title{Clear cached functions}
\usage{
.pb_cache_clear()
}
\value{
invisible: TRUE on success
}
\description{
This function clears the cache for memoised piggyback functions.
}
\examples{
.pb_cache_clear()

}
\keyword{internal}
