% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{gh_new_release}
\alias{gh_new_release}
\title{Create a new release on GitHub repo}
\usage{
gh_new_release(repo, tag, commit = "master", name = tag,
  body = "Data release", draft = FALSE, prerelease = FALSE,
  .token = get_token())
}
\description{
Create a new release on GitHub repo
}
