% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_read.R
\name{guess_read_function}
\alias{guess_read_function}
\title{Guess read function from file extension}
\usage{
guess_read_function(file)
}
\arguments{
\item{file}{filename to parse}
}
\value{
function for reading the file, if found
}
\description{
This function accepts a filename and tries to return a valid function for
reading it.
}
\details{
\code{guess_read_function} understands the following file extensions:
\itemize{
\item rds with \code{readRDS}
\item csv, csv.gz, csv.xz with \code{utils::read.csv}
\item tsv, tsv.gz, tsv.xz with \code{utils::read.delim}
\item parquet with \code{arrow::read_parquet}
\item txt, txt.gz, txt.xz with \code{readLines}
\item json, json.gz, json.xz with \code{jsonlite::fromJSON}
}
}
\seealso{
Other pb_rw: 
\code{\link{guess_write_function}()},
\code{\link{pb_read}()},
\code{\link{pb_write}()}
}
\concept{pb_rw}
\keyword{internal}
