% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_write.R
\name{guess_write_function}
\alias{guess_write_function}
\title{Guess write function from file extension}
\usage{
guess_write_function(file)
}
\arguments{
\item{file}{filename to parse}
}
\value{
function for reading the file, if found
}
\description{
This function accepts a filename and tries to return a valid function for
writing to it.
}
\details{
\code{guess_write_function} understands the following file extensions:
\itemize{
\item rds with \code{saveRDS}
\item csv, csv.gz, csv.xz with \code{utils::write.csv}
\item tsv, tsv.gz, tsv.xz with a modified \code{utils::write.csv} where sep is set to \code{"\\t"}
\item parquet with \code{arrow::write_parquet}
\item txt, txt.gz, txt.xz with \code{writeLines}
\item json, json.gz, json.xz with \code{jsonlite::write_json}
}
}
\seealso{
Other pb_rw: 
\code{\link{guess_read_function}()},
\code{\link{pb_read}()},
\code{\link{pb_write}()}
}
\concept{pb_rw}
\keyword{internal}
