% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{pb_delete}
\alias{pb_delete}
\title{Delete an asset attached to a release}
\usage{
pb_delete(repo, tag = "latest", file)
}
\arguments{
\item{repo}{Repository name in format "owner/repo".}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{file}{path to file to be uploaded}
}
\value{
\code{TRUE} (invisibly) if a file is found and deleted.
Otherwise, returns \code{NULL} (invisibly) if no file matching the name was found.
}
\description{
Delete an asset attached to a release
}
