% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{pb_upload}
\alias{pb_upload}
\title{Upload data to an existing release}
\usage{
pb_upload(repo, tag, file, name = NULL, overwrite = FALSE,
  .token = get_token())
}
\arguments{
\item{repo}{Repository name in format "owner/repo".}

\item{tag}{tag for the release to which this data should be attached.}

\item{file}{path to file to be uploaded}

\item{name}{name for uploaded file. If not provided will use the basename of
`file` (i.e. filename without directory)}

\item{overwrite}{overwrite any existing file with the same name already
attached to the on release?}

\item{.token}{GitHub authentication token}
}
\description{
NOTE: you must first create a release if one does not already exists.
}
\examples{
\dontrun{
# Needs your real token to run
Sys.setenv(GITHUB_TOKEN = "xxxxx")

readr::write_tsv(mtcars,"mtcars.tsv.xz")
ghdata_upload("cboettig/piggyback", "v0.0.3", "mtcars.tsv.xz")
}
}
