% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_download_url.R
\name{pb_download_url}
\alias{pb_download_url}
\title{Get the download url of a given file}
\usage{
pb_download_url(
  file = NULL,
  repo = guess_repo(),
  tag = "latest",
  .token = gh::gh_token()
)
}
\arguments{
\item{file}{character: vector of names of files to be downloaded. If \code{NULL},
all assets attached to the release will be downloaded.}

\item{repo}{string: GH repository name in format "owner/repo". Default \code{guess_repo()}
tries to guess based on current working directory's git repository}

\item{tag}{string: tag for the GH release, defaults to "latest"}

\item{.token}{GitHub authentication token, see \code{\link[gh:gh_token]{gh::gh_token()}}}
}
\value{
the URL to download a file
}
\description{
Returns the URL download for a public file. This can be useful when writing
scripts that may want to download the file directly without introducing any
dependency on \code{piggyback} or authentication steps.
}
\examples{
\dontrun{

pb_download_url("iris.tsv.xz",
                repo = "cboettig/piggyback-tests",
                tag = "v0.0.1")

}
}
