% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_download_url.R
\name{pb_download_url}
\alias{pb_download_url}
\title{Get the download url of a given file}
\usage{
pb_download_url(file = NULL, repo = guess_repo(), tag = "latest",
  .token = get_token())
}
\arguments{
\item{file}{name or vector of names of files to be downloaded. If \code{NULL},
all assets attached to the release will be downloaded.}

\item{repo}{Repository name in format "owner/repo". Will guess
the current repo if not specified.}

\item{tag}{tag for the GitHub release to which this data is attached}

\item{.token}{GitHub authentication token. Typically set from an
environmental variable, e.g. in a \code{.Renviron} file or with
\code{Sys.setenv(GITHUB_TOKEN = "xxxxx")}, which helps prevent
accidental disclosure of a secret token when sharing scripts.}
}
\value{
the URL to download a file
}
\description{
Returns the URL download for a public file. This can be useful when writing
scripts that may want to download the file directly without introducing any
dependency on \code{piggyback} or authentication steps.
}
\examples{
\donttest{

pb_download_url("data/iris.tsv.xz",
                repo = "cboettig/piggyback-tests",
                tag = "v0.0.1")

}
}
