% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{pb_list}
\alias{pb_list}
\title{List all assets attached to a release}
\usage{
pb_list(repo = guess_repo(), tag = "latest", ignore = "manifest.json")
}
\arguments{
\item{repo}{Repository name in format "owner/repo". Will guess
the current repo if not specified.}

\item{tag}{tag for the GitHub release to which this data is attached}

\item{ignore}{a list of files to ignore (if downloading "all" because \code{file=NULL}).
by default, "manifest.json" is ignored as this file is created and used only for
avoiding redundant file transfer by \code{pb_push} and \code{pb_pull}}
}
\value{
a character vector of release asset names, (normalized to local paths)
}
\description{
List all assets attached to a release
}
\details{
To preserve path information, local path delimeters are converted to \code{.2f}
when files are uploaded as assets.  Listing will display the local filename,
with asset names converting the \code{.2f} escape code back to the system delimiter.
}
