% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push-pull.R
\name{pb_push}
\alias{pb_push}
\title{Push data to GitHub}
\usage{
pb_push(tag = "latest", overwrite = TRUE, manifest = ".manifest.json",
  .repo = guess_repo())
}
\arguments{
\item{tag}{name of release/tag on GitHub to which data assets will be
attached. Default is to use the latest available release.}

\item{overwrite}{should existing files be overwritten when hashes do
not match? default \code{TRUE}.}

\item{manifest}{name of the local manifest file. Note: A leading dot
(i.e. indicating a hidden file) in the manifest name will be removed
from the name used on the GitHub asset list.}

\item{.repo}{Name of the repo on GitHub (\code{owner/repo}, i.e.
\code{cboettig/piggyback}). By default will guess the current repository's
GitHub \code{origin}.}
}
\description{
Push all currently tracked data files to GitHub.  Only files identical
to those already on GitHub (by md5sum hash) will not be transferred.
Otherwise, \strong{assumes local version should overwrite existing GitHub
version.} Create a new release if you do not want to overwrite previous
GitHub versions when pushing.
}
\details{
Will only upload tracked files, as identified by the local
manifest.  Add files to tracking with \code{\link{pb_track}} first.
}
\examples{
\dontrun{
pb_push()
}
}
