% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{pb_upload}
\alias{pb_upload}
\title{Upload data to an existing release}
\usage{
pb_upload(file, repo = guess_repo(), tag = "latest", name = NULL,
  overwrite = FALSE, .token = get_token())
}
\arguments{
\item{file}{path to file to be uploaded}

\item{repo}{Repository name in format "owner/repo". Will gess the current
repo if not specified.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{name}{name for uploaded file. If not provided will use the basename of
\code{file} (i.e. filename without directory)}

\item{overwrite}{overwrite any existing file with the same name already
attached to the on release?}

\item{.token}{GitHub authentication token. Typically set from an environmental
variable, e.g. \code{Sys.setenv(GITHUB_TOKEN = "xxxxx")}, which helps prevent
accidental disclosure of a secret token when sharing scripts.}
}
\description{
NOTE: you must first create a release if one does not already exists.
}
\examples{
\dontrun{
# Needs your real token to run
Sys.setenv(GITHUB_TOKEN = "xxxxx")

readr::write_tsv(mtcars,"mtcars.tsv.xz")
ghdata_upload("cboettig/piggyback", "v0.0.3", "mtcars.tsv.xz")
}
}
